+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - ga() is now faster.
 - Implement dfbetas().
 - Separate implementations of dffit() and dffits()
+- Implement diagnose()
 
 
 # v0.9.2